move the json converters into the core library
this allows us to remove the dependency on electionGuard.Utils for ElectionSetup
