+++ bindings/netstandard/ElectionGuard/ElectionGuard.Decryption/Challenge/BallotChallenge.cs
@@ -3,7 +3,6 @@ using ElectionGuard.Decryption.Accumulation;
 using ElectionGuard.ElectionSetup;
 using ElectionGuard.ElectionSetup.Extensions;
 using ElectionGuard.Guardians;
-using Newtonsoft.Json;
 
 namespace ElectionGuard.Decryption.Challenge;
