+++ bindings/netstandard/ElectionGuard/ElectionGuard.Decryption/Challenge/TallyChallenge.cs
@@ -2,7 +2,6 @@
 using ElectionGuard.ElectionSetup;
 using ElectionGuard.ElectionSetup.Extensions;
 using ElectionGuard.Guardians;
-using Newtonsoft.Json;
 
 namespace ElectionGuard.Decryption.Challenge;
