+++ bindings/netstandard/ElectionGuard/ElectionGuard.Decryption/ChallengeResponse/GuardianChallengeResponse.cs
@@ -4,7 +4,6 @@ using ElectionGuard.Decryption.Tally;
 using ElectionGuard.ElectionSetup;
 using ElectionGuard.Extensions;
 using ElectionGuard.Guardians;
-using Newtonsoft.Json;
 
 namespace ElectionGuard.Decryption.ChallengeResponse;
