+++ bindings/netstandard/ElectionGuard/ElectionGuard.Decryption/GlobalUsings.cs
@@ -1,0 +1,1 @@
+global using Newtonsoft.Json;
