+++ bindings/netstandard/ElectionGuard/ElectionGuard.Decryption/Shares/TallyShare.cs
@@ -2,7 +2,6 @@
 using ElectionGuard.ElectionSetup;
 using ElectionGuard.ElectionSetup.Extensions;
 using ElectionGuard.Guardians;
-using Newtonsoft.Json;
 
 namespace ElectionGuard.Decryption.Shares;
