+++ bindings/netstandard/ElectionGuard/ElectionGuard.Encryption/Converters/CiphertextBallotConverter.cs
@@ -4,24 +4,24 @@ using System;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace ElectionGuard.Encryption.Utils.Converters
+namespace ElectionGuard.Converters
 {
-    public class ManifestConverter : JsonConverter
+    public class CiphertextBallotConverter : JsonConverter
     {
         public override bool CanConvert(Type objectType)
         {
-            return objectType == typeof(Manifest);
+            return objectType == typeof(CiphertextBallot);
         }
 
         public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
         {
             var obj = JObject.Load(reader);
-            return new Manifest(obj.ToString());
+            return new CiphertextBallot(obj.ToString());
         }
 
         public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
         {
-            var obj = (Manifest)value;
+            var obj = (CiphertextBallot)value;
             var json = obj.ToJson();
             writer.WriteRawValue(json);
         }
