+++ bindings/netstandard/ElectionGuard/ElectionGuard.Encryption/Converters/CiphertextElectionContextConverter.cs
@@ -4,24 +4,24 @@ using System;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace ElectionGuard.Encryption.Utils.Converters
+namespace ElectionGuard.Converters
 {
-    public class CiphertextBallotConverter : JsonConverter
+    public class CiphertextElectionContextConverter : JsonConverter
     {
         public override bool CanConvert(Type objectType)
         {
-            return objectType == typeof(CiphertextBallot);
+            return objectType == typeof(CiphertextElectionContext);
         }
 
         public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
         {
             var obj = JObject.Load(reader);
-            return new CiphertextBallot(obj.ToString());
+            return new CiphertextElectionContext(obj.ToString());
         }
 
         public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
         {
-            var obj = (CiphertextBallot)value;
+            var obj = (CiphertextElectionContext)value;
             var json = obj.ToJson();
             writer.WriteRawValue(json);
         }
