+++ bindings/netstandard/ElectionGuard/ElectionGuard.Encryption/Converters/ElGamalCiphertextConverter.cs
@@ -2,7 +2,7 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace ElectionGuard.Encryption.Utils.Converters
+namespace ElectionGuard.Converters
 {
     public class ElGamalCiphertextConverter : JsonConverter<ElGamalCiphertext>
     {
