+++ bindings/netstandard/ElectionGuard/ElectionGuard.Encryption/Converters/InternalManifestConverter.cs
@@ -4,24 +4,24 @@ using System;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace ElectionGuard.Encryption.Utils.Converters
+namespace ElectionGuard.Converters
 {
-    public class PlaintextBallotConverter : JsonConverter
+    public class InternalManifestConverter : JsonConverter
     {
         public override bool CanConvert(Type objectType)
         {
-            return objectType == typeof(PlaintextBallot);
+            return objectType == typeof(InternalManifest);
         }
 
         public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
         {
             var obj = JObject.Load(reader);
-            return new PlaintextBallot(obj.ToString());
+            return new InternalManifest(obj.ToString());
         }
 
         public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
         {
-            var obj = (PlaintextBallot)value;
+            var obj = (InternalManifest)value;
             var json = obj.ToJson();
             writer.WriteRawValue(json);
         }
