+++ bindings/netstandard/ElectionGuard/ElectionGuard.Encryption/Converters/ManifestConverter.cs
@@ -4,24 +4,24 @@ using System;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace ElectionGuard.Encryption.Utils.Converters
+namespace ElectionGuard.Converters
 {
-    public class InternalManifestConverter : JsonConverter
+    public class ManifestConverter : JsonConverter
     {
         public override bool CanConvert(Type objectType)
         {
-            return objectType == typeof(InternalManifest);
+            return objectType == typeof(Manifest);
         }
 
         public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
         {
             var obj = JObject.Load(reader);
-            return new InternalManifest(obj.ToString());
+            return new Manifest(obj.ToString());
         }
 
         public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
         {
-            var obj = (InternalManifest)value;
+            var obj = (Manifest)value;
             var json = obj.ToJson();
             writer.WriteRawValue(json);
         }
