+++ bindings/netstandard/ElectionGuard/ElectionGuard.Encryption/Converters/PlaintextBallotConverter.cs
@@ -4,24 +4,24 @@ using System;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
-namespace ElectionGuard.Encryption.Utils.Converters
+namespace ElectionGuard.Converters
 {
-    public class CiphertextElectionContextConverter : JsonConverter
+    public class PlaintextBallotConverter : JsonConverter
     {
         public override bool CanConvert(Type objectType)
         {
-            return objectType == typeof(CiphertextElectionContext);
+            return objectType == typeof(PlaintextBallot);
         }
 
         public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
         {
             var obj = JObject.Load(reader);
-            return new CiphertextElectionContext(obj.ToString());
+            return new PlaintextBallot(obj.ToString());
         }
 
         public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
         {
-            var obj = (CiphertextElectionContext)value;
+            var obj = (PlaintextBallot)value;
             var json = obj.ToJson();
             writer.WriteRawValue(json);
         }
